plugins/chainlit-pilot/requirements.txt
chainlit>=1.3.0
openai>=1.0.0
PyJWT>=2.8.0
